hw-4: refactor dish-counter & progress-bar
