+++ src/components/scroll-progress-bar/scroll-progress-bar.jsx
@@ -1,28 +1,7 @@
-import { useEffect, useState } from "react";
+import { useProgress } from "./use-progress";
 
 export const ScrollProgressBar = () => {
-    const [progressBarWidth, setProgressBarWidth] = useState("0");
-
-    const getScrollPercent = () => {
-        const scrollTop = document.documentElement.scrollTop; // px
-        const windowHeight =
-            document.documentElement.scrollHeight -
-            document.documentElement.clientHeight;
-        const scrolledPercent = scrollTop / windowHeight * 100 + "%";
-
-        return (
-            scrolledPercent
-        )
-    };
-
-    useEffect(() => {
-        const scrollCallback = () => {
-            setProgressBarWidth(getScrollPercent);
-        };
-
-        window.addEventListener("scroll", scrollCallback);
-        return () => window.removeEventListener("scroll", scrollCallback);
-    }, []);
+    const progress = useProgress();
 
     const stylesProgressBar = {
         height: "10px",
@@ -30,7 +9,7 @@ export const ScrollProgressBar = () => {
         position: "fixed",
         top: 0,
         left: 0,
-        width: progressBarWidth,
+        width: progress,
     };
     return (
         <div style={stylesProgressBar}></div>
